refactor: move disclaimer component
