+++ src/app/tech-carbon-estimator/tech-carbon-estimator.component.ts
@@ -15,7 +15,7 @@ import { CarbonEstimation, EstimatorValues } from '../types/carbon-estimator';
 import { CarbonEstimationService } from '../services/carbon-estimation.service';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { DisclaimerComponent } from '../disclaimer/disclaimer.component';
+import { DisclaimerComponent } from '../components/disclaimer/disclaimer.component';
 import { TabsComponent } from '../tab/tabs/tabs.component';
 import { TabItemComponent } from '../tab/tab-item/tab-item.component';
 import { ExportModal } from '../export-modal/export-modal.component';
